Move alert contact data sources into domain package
