+++ internal/provider/provider.go
@@ -13,6 +13,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-framework/resource"
 	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/uptimerobot/terraform-provider-uptimerobot/internal/client"
+	"github.com/uptimerobot/terraform-provider-uptimerobot/internal/provider/alertcontact"
 	"github.com/uptimerobot/terraform-provider-uptimerobot/internal/provider/currentuser"
 	"github.com/uptimerobot/terraform-provider-uptimerobot/internal/provider/iprange"
 	"github.com/uptimerobot/terraform-provider-uptimerobot/internal/provider/tag"
@@ -114,9 +115,9 @@ func (p *UptimeRobotProvider) Resources(ctx context.Context) []func() resource.R
 
 func (p *UptimeRobotProvider) DataSources(ctx context.Context) []func() datasource.DataSource {
 	return []func() datasource.DataSource{
-		NewAllAlertContactsDataSource,
-		NewAlertContactDataSource,
-		NewAlertContactsDataSource,
+		alertcontact.NewAllDataSource,
+		alertcontact.NewDataSource,
+		alertcontact.NewListDataSource,
 		currentuser.NewDataSource,
 		iprange.NewDataSource,
 		NewIntegrationDataSource,
